+++ src/hotspot/share/gc/z/zPhysicalMemoryManager.cpp
@@ -104,7 +104,7 @@ void ZPhysicalMemoryManager::try_enable_uncommit(size_t min_capacity, size_t max
   // Test if uncommit is supported by the operating system by committing
   // and then uncommitting a granule.
   const zbacking_index offset{};
-  if (!commit(&offset, ZGranuleSize, -1) || !uncommit(&offset, ZGranuleSize)) {
+  if (!commit(&offset, ZGranuleSize, (uint32_t)-1) || !uncommit(&offset, ZGranuleSize)) {
     log_info_p(gc, init)("Uncommit: Implicitly Disabled (Not supported by operating system)");
     FLAG_SET_ERGO(ZUncommit, false);
     return;
